Run capabilities: imports refactoring
