+++ client/src/components/pipelines/launch/form/utilities/parameters.js
@@ -71,3 +71,23 @@ export const reservedParameters = [
   CP_CAP_REQUESTS_GPU,
   CP_CAP_REQUESTS_RAM
 ];
+
+export const RUN_CAPABILITIES = {
+  dinD: 'DinD',
+  singularity: 'Singularity',
+  systemD: 'SystemD',
+  noMachine: 'NoMachine',
+  module: 'Module',
+  disableHyperThreading: 'Disable Hyper-Threading',
+  dcv: 'NICE DCV'
+};
+
+export const RUN_CAPABILITIES_PARAMETERS = {
+  [RUN_CAPABILITIES.dinD]: CP_CAP_DIND_CONTAINER,
+  [RUN_CAPABILITIES.singularity]: CP_CAP_SINGULARITY,
+  [RUN_CAPABILITIES.systemD]: CP_CAP_SYSTEMD_CONTAINER,
+  [RUN_CAPABILITIES.noMachine]: CP_CAP_DESKTOP_NM,
+  [RUN_CAPABILITIES.module]: CP_CAP_MODULES,
+  [RUN_CAPABILITIES.disableHyperThreading]: CP_DISABLE_HYPER_THREADING,
+  [RUN_CAPABILITIES.dcv]: CP_CAP_DCV
+};
